+++ Serwer/SerwerApplication/Vector2D.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-Vector2D::Vector2D(float x, float y)
+Vector2D::Vector2D(double x, double y)
 	:x(x), y(y)
 {
 }
@@ -9,28 +9,28 @@ Vector2D::~Vector2D()
 {
 }
 
-void Vector2D::SetX(float x)
+void Vector2D::SetX(double x)
 {
 	this->x = x;
 }
 
-void Vector2D::SetY(float y)
+void Vector2D::SetY(double y)
 {
 	this->y = y;
 }
 
-void Vector2D::SetVector(float x, float y)
+void Vector2D::SetVector(double x, double y)
 {
 	this->x = x;
 	this->y = y;
 }
 
-float Vector2D::getX() const
+double Vector2D::getX() const
 {
 	return this->x;
 }
 
-float Vector2D::getY() const
+double Vector2D::getY() const
 {
 	return this->y;
 }
@@ -40,17 +40,17 @@ bool Vector2D::operator!=(const Vector2D& vector) const
 	return (this->x != vector.x || this->y != vector.y);
 }
 
-float Vector2D::length() const
+double Vector2D::length() const
 {
 	return sqrt(this->x * this->x + this->y * this->y);
 }
 
-Vector2D Vector2D::operator*(const float& scalar) const
+Vector2D Vector2D::operator*(const double& scalar) const
 {
 	return Vector2D(this->x * scalar, this->y * scalar);
 }
 
-Vector2D Vector2D::operator/(const float& scalar) const
+Vector2D Vector2D::operator/(const double& scalar) const
 {
 	if(scalar==0) {
 		return  Vector2D(0, 0);
@@ -108,12 +108,12 @@ bool Vector2D::operator>=(const Vector2D& vector) const
 	return (this->length() >= vector.length());
 }
 
-float Vector2D::Dot(const Vector2D& a, const Vector2D& b)
+double Vector2D::Dot(const Vector2D& a, const Vector2D& b)
 {
 	return ((a.x * b.x) + (a.y * b.y));
 }
 
-float Vector2D::CrossProduct(const Vector2D& a, const Vector2D& b)
+double Vector2D::CrossProduct(const Vector2D& a, const Vector2D& b)
 {
 	return ((a.x * b.y) - (a.y * b.x));
 }
